#56 error correction

+++ src/atmega2560p/hal/analog.rs
@@ -563,6 +563,6 @@ pub fn analog_reference(reftype: RefType) {
     }
 }
 ///This function converts output generated from analog_read() in form to be used as input in analog_write().
-pub fn map_from1023_to255(val: u32) -> u32 {
-    255 * (val / 1023)
+pub fn map_from1023_to255(val: u32) -> u8 {
+    255 * (val / 1023) as u8
 }
